Add files via upload
The goal of this project is to primarily focus on tracking supply stock in the warehouse to monitor and plan for future consumption projections.

It is based on design principles and processes according to the relevant warehouse business operation system. In some cases, by connecting to the working system, the multi- fact table can be transformed for in-depth calculation and analysis in SAP S4 HANA version .

Thank you for your review and comments for further development.